# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación de la API de Reserva de Espacios",
        "description": "API para la gestión de reservas de espacios como salas de reuniones, auditorios, y más.",
        "contact": {
            "email": "rmartinez4298@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/filter-places": {
            "get": {
                "tags": [
                    "Places"
                ],
                "summary": "Filter places based on criteria",
                "description": "Filter places based on type, capacity, and availability dates and times.",
                "operationId": "filterPlaces",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Space type (Ej: oficina, sala de reuniones, etc.)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "office"
                        }
                    },
                    {
                        "name": "capacity",
                        "in": "query",
                        "description": "Minimum capacity of the space",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Start date for availability check",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-10-15"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "End date for availability check",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-10-16"
                        }
                    },
                    {
                        "name": "start_time",
                        "in": "query",
                        "description": "Start time for availability check (HH:mm)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "time",
                            "example": "09:00"
                        }
                    },
                    {
                        "name": "end_time",
                        "in": "query",
                        "description": "End time for availability check (HH:mm)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "time",
                            "example": "17:00"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Places filtered successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "places": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Place"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "500": {
                        "description": "Server error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Server error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Place": {
                "required": [
                    "id",
                    "name",
                    "description",
                    "capacity",
                    "default_days",
                    "default_hours"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Sala de Reuniones A"
                    },
                    "description": {
                        "type": "string",
                        "example": "Una sala espaciosa para reuniones."
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 10
                    },
                    "default_days": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "Monday"
                        }
                    },
                    "default_hours": {
                        "type": "string",
                        "example": "09:00-17:00"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}